requirements.txt
PySide6>=6.6.0
sounddevice>=0.4.6
numpy>=1.24.0
keyboard>=0.13.5
pyperclip>=1.8.2
PyYAML>=6.0
pywin32>=306
faster-whisper>=1.0.0
